Fixed bug where thread stack kept growing, fixed manager bug
Implemented hailstone sequence with yielding as a more complex example,
which led to discovering two bugs. The first was a realloc bug when
we wanted to increase the size of the array that holds threads in it
in the thread manager. Easy. The second was a bug where the t_StackCur
variable for each thread would grow by four bytes after every yield,
caused by trying to account for an argument to the function when
calculating where esp leaves off, when in fact it shouldn't be
accounted for as the caller handles arguments it pushes to the
callee, not the callee itself.